pyproject.toml
[tool.poetry]
name = "minitools"
version = "0.1.0"
description = ""
authors = ["Methuselah Nwodobeh <methuselah.nwodobeh@amalitech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.4"
django = "^4.2.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.6"
django-cors-headers = "^4.0.0"
django-extensions = "^3.2.3"
django-decouple = "^2.1"
djangorestframework-simplejwt = "^5.2.2"
celery = "^5.3.0"
redis = "^4.5.5"
django-silk = "^5.0.3"
django-prometheus = "^2.3.1"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
flower = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
